requirements.txt
Flask==2.0.3
line-bot-sdk==2.4.2
gunicorn==20.1.0